Throw a fatal error when 'strainer test' is called on a chef_repo with no arguments

+++ lib/strainer/runner.rb
@@ -65,7 +65,9 @@ module Strainer
       end
 
       # Move the logfile back over
-      FileUtils.mv(Strainer.logfile_path, Strainer.sandbox_path.join('strainer.out'))
+      if File.exist?(Strainer.sandbox_path.join('strainer.out'))
+        FileUtils.mv(Strainer.logfile_path, Strainer.sandbox_path.join('strainer.out'))
+      end
 
       if @report.values.collect(&:values).flatten.all?
         Strainer.ui.say "Strainer marked build OK"
